requirements.txt
# 小红书博主笔记采集 API 服务依赖
# Python >= 3.9

# Web 框架
fastapi==0.109.2
uvicorn[standard]==0.27.1

# HTTP 异步请求
httpx==0.27.0

# JS 签名执行（抖音 a_bogus）
PyExecJS==1.5.1

# 数据验证
pydantic==2.6.1
pydantic-settings==2.2.1

# 环境变量
python-dotenv==1.0.1


